extras/serialization/daglite-serialization/pyproject.toml
# --- build system ---

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

# --- project & dependencies ---

[project]
name = "daglite-serialization"
description = "Serialization plugin for daglite with support for numpy, pandas, and PIL."
dynamic = ["version"]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "daglite>=0.4.0",
]

[project.optional-dependencies]
numpy = ["numpy>=1.20"]
pandas = ["pandas>=1.0", "pyarrow>=10.0"]
pillow = ["pillow>=8.0"]
all = [
    "numpy>=1.20",
    "pandas>=1.0",
    "pyarrow>=10.0",
    "pillow>=8.0",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-cov",
    "ruff",
]
test = ["pytest", "pytest-cov", "pytest-mock"]

# --- tool.setuptools ---

[tool.setuptools.dynamic.version]
attr = "daglite_serialization.__version__"

[tool.setuptools.packages.find]
where = ["src"]

# --- tool.pytest ---

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --strict-config --import-mode importlib"
pythonpath = ["."]
testpaths = ["tests"]

# --- tool.ruff ---

[tool.ruff]
line-length = 100
show-fixes = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "D",    # pydocstyle errors
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]
"test*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"
